requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
plotly==5.18.0
python-dotenv==1.0.0
openai>=1.0.0
chardet
fastapi
pdfplumber
pydantic
python-magic
uvicorn
numpy
openai
pandas
plotly
python-dotenv
streamlit
jinja2
weasyprint
jinja2
weasyprint
